+++ bdsim.cc
@@ -1,3 +1,21 @@
+/* 
+Beam Delivery Simulation (BDSIM) Copyright (C) Royal Holloway, 
+University of London 2001 - 2017.
+
+This file is part of BDSIM.
+
+BDSIM is free software: you can redistribute it and/or modify 
+it under the terms of the GNU General Public License as published 
+by the Free Software Foundation version 3 of the License.
+
+BDSIM is distributed in the hope that it will be useful, but 
+WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
+*/
 /**
  * @file bdsim.cc
  *
@@ -47,6 +65,7 @@
 #include "BDSSDManager.hh"
 #include "BDSSteppingAction.hh"
 #include "BDSStackingAction.hh"
+#include "BDSTemporaryFiles.hh"
 #include "BDSTrackingAction.hh"
 #include "BDSUtilities.hh"
 #include "BDSVisManager.hh"
@@ -289,6 +308,7 @@ int main(int argc,char** argv)
   G4cout << __FUNCTION__ << "> instances deleting..."<<G4endl;
 #endif
   delete BDSAcceleratorModel::Instance();
+  delete BDSTemporaryFiles::Instance();
   delete globalConstants;
   delete BDSMaterials::Instance();