+++ src/BDSExecOptions.cc
@@ -1,3 +1,21 @@
+/* 
+Beam Delivery Simulation (BDSIM) Copyright (C) Royal Holloway, 
+University of London 2001 - 2017.
+
+This file is part of BDSIM.
+
+BDSIM is free software: you can redistribute it and/or modify 
+it under the terms of the GNU General Public License as published 
+by the Free Software Foundation version 3 of the License.
+
+BDSIM is distributed in the hope that it will be useful, but 
+WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
+*/
 #include "BDSExecOptions.hh"
 
 #include <cstdlib>
@@ -22,11 +40,14 @@ BDSExecOptions::BDSExecOptions(int argc, char **argv):
   ignoreSIGINT(false)
 {
   Parse(argc, argv);
-  /// after parsing the absolute path can be reconstructed  
-  options.set_value("bdsimPath", std::string(GetPath(options.inputFileName)));
+
+  /// after parsing the absolute path can be reconstructed
+  std::string bp = GetPath(options.inputFileName);
+  options.set_value("bdsimPath", bp);
 #ifdef BDSDEBUG
   G4cout << __METHOD_NAME__ << "BDSIMPATH set to: " << options.bdsimPath << G4endl;
 #endif
+
   if (options.recreate)
     {
       BDSOutputLoader loader(options.recreateFileName);
@@ -343,10 +364,14 @@ G4String BDSExecOptions::GetPath(G4String fileName)
 	  fullPath = inputFilepath;
 	}
       else
-	{fullPath = BDS::GetCurrentDir() + "/" + inputFilepath;}
+	{
+	  G4String curDir = BDS::GetCurrentDir();
+	  fullPath = curDir + "/" + inputFilepath;
+	}
     }
-  // add additional slash just to be safe
-  fullPath += "/";
+  
+  if (fullPath.back() != '/') // ensure ends in '/'
+    {fullPath += "/";} // only add if needed
 #ifdef BDSDEBUG
   G4cout << __METHOD_NAME__ << "fullpath = " << fullPath << G4endl;
 #endif