+++ src/obsolete/BDSPlanckScatter.cc
@@ -1,21 +1,21 @@
-/* BDSIM code.    Version 1.0
-   Author: Grahame A. Blair, Royal Holloway, Univ. of London.
-   Last modified 24.7.2002
-   Copyright (c) 2002 by G.A.Blair.  ALL RIGHTS RESERVED. 
+/* 
+Beam Delivery Simulation (BDSIM) Copyright (C) Royal Holloway, 
+University of London 2001 - 2017.
+
+This file is part of BDSIM.
+
+BDSIM is free software: you can redistribute it and/or modify 
+it under the terms of the GNU General Public License as published 
+by the Free Software Foundation version 3 of the License.
+
+BDSIM is distributed in the hope that it will be useful, but 
+WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 */
-//      ------------ BDSPlanckScatter physics process --------
-//                     by Grahame Blair, 18 October 2001
-#include "BDSGlobalConstants.hh" 
-
-#include "BDSPlanckScatter.hh"
-#include "G4ios.hh"
-#include "G4Gamma.hh"
-
-BDSPlanckScatter::BDSPlanckScatter():G4VEnergyLossProcess("PlanckScatt")
-{
-
-  // TODO: change to appropriate definition!!!
-  itsTemperature = 300 * CLHEP::kelvin;
 
   if(itsTemperature<=0.){G4Exception("BDSPlanckScatter: Invalid Temperature", "-1", FatalException, "");}
   itsPlanckEngine=new BDSPlanckEngine(itsTemperature);